requirements.txt
boto3
Pillow
rembg[cpu,cli]
numpy<2